foodgram/requirements.txt
requests==2.28.2
django==2.2.19
psycopg2-binary==2.8.6
python-dotenv==0.21.1
pillow==9.4.0
sorl-thumbnail==12.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
django-filter==21.1
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.6